HyperGarageSale
HyperGarageSale is an Android app to sell products. The user will be
able to add a new item to sell along with a picture and location to the
ad.  When the ad is posted the app will also say a thank you note for
adding the ad.The main purpose of this app is used to play around with
the following APIs:
1.Camera API
2.Location API
3.Content Providers
4.Text-To-Speech API
5.ReyclerView